pyproject.toml
[project]
name = "fcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chardet<6",
    "fastmcp>=3.1.1",
    "playwright>=1.58.0",
    "readability-lxml>=0.8.4.1",
    "requests>=2.32.5",
]
